+++ kmpnotifier-core/src/wasmJsMain/kotlin/com/mmk/kmpnotifier/di/PlatformModule.web.kt
@@ -2,12 +2,7 @@ package com.mmk.kmpnotifier.di
 
 import com.mmk.kmpnotifier.permission.PermissionUtil
 import com.mmk.kmpnotifier.permission.WebPermissionUtilImpl
-import org.koin.core.module.dsl.factoryOf
-import org.koin.dsl.bind
-import org.koin.dsl.module
 
+internal actual val platform: Platform = Platform.Web
 
-internal actual val platformModule = module {
-    factory { Platform.Web } bind Platform::class
-    factoryOf(::WebPermissionUtilImpl) bind PermissionUtil::class
-}
+internal actual fun createPermissionUtil(): PermissionUtil = WebPermissionUtilImpl()